fix(payment): widen amountVnd int -> long (VND overflow)
Order.AmountVnd and OrderResponse.AmountVnd were int (cap ~2.147e9), so a
large credit pack or a gathered postpaid invoice could exceed ~2.1B VND and
overflow silently, corrupting the stored/returned amount and the PayOS
charge. payment.md already specifies amountVnd: long / amount_vnd bigint —
the code was the outlier.

Fix (preferred path): both properties become long, plus an additive EF
migration AmountVndToBigint that alters orders.amount_vnd integer -> bigint
(reversible Down). It does not edit the squashed InitialCreate and is NOT
applied here (repo policy: schema applied manually before deploy). No
payment.md change needed (already long/bigint). Adds a round-trip test for
a VND value beyond int.MaxValue.

+++ src/services/Isas.PaymentService/DTOs/OrderRequest.cs
@@ -26,7 +26,7 @@ namespace Isas.PaymentService.DTOs
             public Guid? PackageId { get; set; }
             public Guid? InvoiceId { get; set; }
             public OrderStatus Status { get; set; }
-            public int AmountVnd { get; set; }
+            public long AmountVnd { get; set; }   // khớp Order.AmountVnd (amount_vnd bigint — payment.md §DB)
             public long PayosOrderCode { get; set; }
             public DateTime ExpiredAt { get; set; }
             public DateTime? PaidAt { get; set; }
